+++ src/Http/Livewire/SeoIssuesTable.php
@@ -12,9 +12,7 @@ class SeoIssuesTable extends Component
     use WithPagination;
 
     public ?int $scanId = null;
-
     public string $severity = '';
-
     public string $type = '';
 
     public function updatingSeverity(): void