+++ src/helpers.php
@@ -1,7 +1,5 @@
 <?php
 
-use Step2dev\LazySeoTools\Services\JsonLdService;
-use Step2dev\LazySeoTools\Services\SchemaService;
 use Step2dev\LazySeoTools\Services\SeoManager;
 
 if (! function_exists('seo')) {
@@ -11,16 +9,17 @@ if (! function_exists('seo')) {
     }
 }
 
+
 if (! function_exists('seo_schema')) {
     function seo_schema(string $type = 'webPage', array $data = []): array
     {
-        return app(SchemaService::class)->make($type, $data);
+        return app(\Step2dev\LazySeoTools\Services\SchemaService::class)->make($type, $data);
     }
 }
 
 if (! function_exists('seo_jsonld')) {
     function seo_jsonld(string $type = 'webPage', array $data = []): string
     {
-        return app(JsonLdService::class)->script($type, $data);
+        return app(\Step2dev\LazySeoTools\Services\JsonLdService::class)->script($type, $data);
     }
 }